changes to a single database repository

+++ mcare-dashboard-web/src/main/java/org/mcare/common/interfaces/DatabaseRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface DatabaseRepository {
 	
 	public <T> long save(T t);
-	public <T> long update(T t);
+	public <T> int update(T t);
 	public <T> boolean delete(T t);
 	
 	public <T> T findById(long id, String fieldName, Class<?> className);
